chore: clear the lint backlog so the gate can be turned on
156 of 159 clippy findings, across the whole workspace. Mostly mechanical —
index loops rewritten as iterator loops, redundant conversions dropped, range
patterns collapsed, `vec![]` for build-then-push, markdown paragraphs
separated from the lists they had been folded into.

Three pieces were not mechanical and are worth naming:

The FFI crate had twenty `unsafe extern "C"` entry points with no safety
contract. They share two invariants — what makes a handle valid, and what
makes a `*const c_char` valid — so those are stated once at module level and
each function names which of its parameters they apply to. The C-string rule
is a real guarantee rather than a caution: NULL and non-UTF-8 are read as the
empty string, because a recorder that loses a name should write a trace with
an empty name in it rather than abort the program it is recording.

Two dead extern declarations turned out to be the name-taking entry points the
`_by_type_id` forms replaced at their call sites. They are kept and marked,
the way this file already keeps `trace_writer_register_variable_raw`: the C
ABI still exports them for callers holding a type name rather than an
interned id.

`build_compat_mmap` and `build_compat_platform` were unused in Rust. They are
the canonical fixtures the Nim compat test serializes against, so each gained
a round-trip test rather than being deleted.

What remains is three `too_many_arguments`, left deliberately: two of them
mirror C ABI entry points one for one, and grouping the parameters would
decouple the two signatures so a mismatch stops being visible at the call
site.

+++ codetracer_trace_writer_nim/tests/column_decode_cross_reader.rs
@@ -234,7 +234,7 @@ fn nim_and_rust_decoders_agree_on_every_step_in_fixture() {
     // with the Rust decoder is unambiguously a Rust bug rather than
     // a writer/reader mismatch.
     for (i, exp) in expected.iter().enumerate() {
-        assert_eq!(nim_path_ids[i] as u64, exp.file as u64, "step {i}: writer→Nim path_id mismatch");
+        assert_eq!(nim_path_ids[i] as u64, exp.file, "step {i}: writer→Nim path_id mismatch");
         assert_eq!(nim_lines[i] as u32, exp.line, "step {i}: writer→Nim line mismatch");
         assert_eq!(nim_columns[i] as u32, exp.column, "step {i}: writer→Nim column mismatch");
     }
@@ -263,7 +263,7 @@ fn nim_and_rust_decoders_agree_on_every_step_in_fixture() {
         // And the Rust decoder MUST agree byte-for-byte with the Nim
         // decoder's answer (the actual cross-reader consistency
         // assertion the M4 plan demands).
-        assert_eq!(u64::from(pos.file), nim_path_ids[i], "step {i}: Rust↔Nim file disagreement at GLI {gli}",);
+        assert_eq!(pos.file, nim_path_ids[i], "step {i}: Rust↔Nim file disagreement at GLI {gli}",);
         assert_eq!(u64::from(pos.line), nim_lines[i], "step {i}: Rust↔Nim line disagreement at GLI {gli}",);
         assert_eq!(
             u64::from(pos.column),
